revamp walltime estimation: now more precise, and less queue-limit dependant.
Instead of trying to guess the "correct" queue, with all its
hardcoded site-specific values, we now calculate only  the expected duration.
The local cluster can figure out which queue that becomes by itself.

A rare case of doing less and the end-result being better :-D

The 13 MB/s value was derived experimentally, and matches closely to values
found scattered on the internet.
Reports of ~100MB/s also appear, but those skip compression, which is not
a can of worms I want to open at the moment.