Rework 0-generate links for better file-exist checking.
Do away with the convoluted awk script that generates a bash script.
Instead, directly do the linking, but only if the file exists.
If the target file doesn't exist, emit a warning.

Log messages are now colored for extra fancyness, and emitted to stderr
where appropriate. (Since it's 2020, we just assume terminal colors work.)

Closes #3 on GitHub.