submit encryption jobs: fix count of already-encrypted files.
When checking which files are already encrypted, we wrongfully read into a bash-array with IFS=''.
This caused all files to be concatenated and counted as a single file
This affected nothing except a user-info output, but it was still sloppy.
Fix it to IFS=\n (line-ending), so that each file is counted individually.